refactor(desktop): split errors.rs into errors/ module (#379)
The 811-line errors.rs exceeded the ~500-line one-concept-per-file
guideline. Split into a directory module with three concerns, no
behavior change — the #[serde(tag = "code", ...)] wire contract (which
the frontend errors.ts union depends on) is byte-identical:

- errors/types.rs    — AppError / AppWarning enum definitions
- errors/mapping.rs  — map_ffi_error + impl From<FfiVaultError>
- errors/tests.rs    — serde round-trip + mapping-routing suite
- errors/mod.rs      — module doc + declarations + re-exports

The test module keeps the name `tests` under `errors` so the
`errors::tests::settings_clamped_warning_carries_both_values`
cross-reference in dtos/manifest.rs stays valid. The public surface
(AppError, AppWarning, map_ffi_error) is unchanged; all 18 downstream
`use crate::errors::…` sites compile untouched.

All 32 errors:: tests + full desktop suite (179+62+18) pass; clippy
-D warnings, cargo fmt --check, and RUSTDOCFLAGS=-D warnings cargo doc
all clean.
